+++ CHANGELOG.md
@@ -55,6 +55,7 @@ This release changes definitions of many API methods.
 - Feat: improve mpd communication
 - Feat: add quick remove and quick play buttons to more views
 - Feat: support MUSICBRAINZ_RELEASEGROUPID tag (MPD 0.24)
+- Feat: handle preflighted requests in CORS
 - Upd: Mongoose 7.10 #1024
 - Upd: Bootstrap 5.3.0 final
 - Upd: libmympdclient 1.0.21